fix: modularize schedule api calls
